+++ internal/js_parser/js_parser_lower.go
@@ -88,11 +88,6 @@ func (p *parser) markSyntaxFeature(feature compat.JSFeature, r logger.Range) (di
 			"Top-level await is not available in %s", where))
 		return
 
-	case compat.ArbitraryModuleNamespaceNames:
-		p.log.AddError(&p.tracker, r, fmt.Sprintf(
-			"Using a string as a module namespace identifier name is not supported in %s", where))
-		return
-
 	case compat.Bigint:
 		// Transforming these will never be supported
 		p.log.AddError(&p.tracker, r, fmt.Sprintf(